backend/requirements.txt
asgiref==3.2.7
Django==3.0.5
pytz==2019.3
sqlparse==0.3.1
django-cors-headers==3.1.1
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0